Extract printFlowMapping, printFlowSequence
